replicate performance for conll_de experiment (disabling correct_bias). add documentation
